virtual-assistant/requirements.txt
absl-py==2.3.1
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altgraph==0.17.2
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.10.0
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
babel==2.17.0
bangla==0.0.5
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
camb-sdk @ git+https://github.com/Camb-ai/cambai-python-sdk@270b02203063108717fdf153aec0001d3b6bd163
catalogue==2.0.10
certifi==2025.8.3
cffi==2.0.0
cfgv==3.4.0
charset-normalizer==3.4.3
click==8.1.8
cloudpathlib==0.22.0
confection==0.1.5
conformer==0.3.2
contourpy==1.2.1
coqpit==0.0.17
csvw==3.7.0
curated-tokenizers==0.0.9
curated-transformers==0.1.1
cycler==0.12.1
cymem==2.0.11
Cython==3.1.3
dateparser==1.1.8
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.29.0
distlib==0.4.0
dlinfo==2.0.0
docopt==0.6.2
einops==0.8.1
elevenlabs==2.15.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
encodec==0.1.1
espeakng-loader==0.2.4
exceptiongroup==1.3.0
filelock==3.19.1
Flask==3.1.2
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
future==0.18.2
gitdb==4.0.12
GitPython==3.1.45
grpcio==1.67.1
grpcio-tools==1.67.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.1.10
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
identify==2.6.15
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
isodate==0.7.2
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.2
jsonlines==1.2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.7
kokoro==0.7.16
langcodes==3.5.0
language-tags==1.2.0
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llvmlite==0.43.0
loguru==0.7.3
macholib==1.15.2
marisa-trie==1.3.1
Markdown==3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
misaki==0.9.4
ml_dtypes==0.5.3
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
murmurhash==1.0.13
mutagen==1.47.0
networkx==2.8.8
nltk==3.9.1
nodeenv==1.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
ollama==0.6.0
onnx==1.19.1
opencv-python==4.10.0.84
packaging==25.0
pandas==1.5.3
phonemizer-fork==3.3.2
pillow==11.3.0
pip-check-reqs==2.5.5
pkuseg==0.0.25
platformdirs==4.4.0
playaudio==1.0.5
polars==1.33.1
pooch==1.8.2
pre_commit==4.3.0
preshed==3.0.10
propcache==0.3.2
protobuf==5.29.5
psutil==7.0.0
pvcobra==2.1.0
pvporcupine==3.0.5
pvrecorder==1.2.7
PyAudio==0.2.14
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pykakasi==2.3.0
pynndescent==0.5.13
pyparsing==3.2.4
pypinyin==0.55.0
pysbd==0.3.4
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
rdflib==7.2.1
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
resemble-perth==1.0.1
rfc3986==1.5.0
rich==14.1.0
rpds-py==0.27.1
s3tokenizer==0.2.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
seaborn==0.13.2
segments==2.3.0
shellingham==1.5.4
simpleaudio==1.0.4
six==1.17.0
smart_open==7.3.1
smmap==5.0.2
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.7
spacy-curated-transformers==0.3.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.14.3
srsly==2.5.1
SudachiDict-core==20250825
SudachiPy==0.6.10
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.1.0
thinc==8.3.4
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.8.0
torch-vision==0.1.6.dev0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.46.3
typeguard==4.4.4
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzlocal==5.3.1
ultralytics==8.3.203
ultralytics-thop==2.0.17
umap-learn==0.5.7
Unidecode==1.4.0
uritemplate==4.2.0
urllib3==2.5.0
virtualenv==20.35.3
wasabi==1.1.3
weasel==0.4.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.20.1
zipp==3.23.0
